requirements.txt
amqp==2.5.2
Babel==2.8.0
billiard==3.6.1.0
cachetools==4.0.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
flower==0.9.3
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.15.0
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
tornado==5.1.1
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
asgiref==3.2.3
Django==3.0.2
sqlparse==0.3.0
psycopg2-binary
